chore:Refactor file and folder names to follow naming conventions
