+++ server/app.js
@@ -2,8 +2,8 @@ import express from 'express'
 import morgan from 'morgan'
 import cors from 'cors'
 import cookieParser from 'cookie-parser'
-import errorHandler from './middlewares/errorHandler.js'
-import userRouter from './routes/userRoutes.js'
+import errorHandler from './middlewares/error.handler.js'
+import userRouter from './routes/user.routes.js'
 import AppError from './utils/appError.js'
 import logger from './utils/logger.js'
