+++ server/controllers/auth.controller.js
@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken'
 import asyncHandler from '../utils/asyncHandler.js'
-import User from '../models/userModel.js'
+import User from '../models/user.model.js'
 import AppError from '../utils/appError.js'
 
 /**
